requirements.txt
aniso8601==8.0.0
astroid==2.4.2
click==7.1.2
Flask==1.1.2
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mysql-connector-python==8.0.21
protobuf==3.12.4
pylint==2.5.3
pytz==2020.1
six==1.15.0
toml==0.10.1
Werkzeug==1.0.1
wrapt==1.12.1
